+++ lib/tempus/tempus_helper.rb
@@ -1,0 +1,7 @@
+module TempusHelper
+  def to_tempus
+    Tempus.new(self)
+  end
+end
+
+Object.send(:include, TempusHelper)
