+++ spec/tempus/tempus_helper_spec.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+RSpec.describe TempusHelper do
+  let(:hours) { Tempus.new(30.hours + 5.minutes + 3.seconds) }
+  let(:negative_hours) { Tempus.new(- 30.hours - 5.minutes - 3.seconds) }
+
+  it 'String to Tempus' do
+    expect('12:05:35'.to_tempus.data).to eq(Tempus.new(12.hours + 5.minutes + 35.seconds).data)
+    expect('01:00:01'.to_tempus.data).to eq(Tempus.new(1.hours + 1.seconds).data)
+    expect('30:5:3'.to_tempus).to eq(hours)
+    expect('-30:5:3'.to_tempus).to eq(negative_hours)
+  end
+
+  it 'Number to Tempus' do
+    expect(3723.to_tempus.to_s).to eq('01:02:03')
+    expect(3723.to_tempus.inspect).to eq('<Tempus seconds=3723.0, formated=01:02:03>')
+  end
+
+  it 'Float to Tempus' do
+    expect(3723.0.to_tempus.to_string).to eq('01:02:03')
+  end
+
+  it 'Time to Tempus' do
+    expect(Time.parse('2014-01-01 01:02:03').to_tempus.to_i).to eq(3723)
+  end
+
+  it 'nil to Tempus' do
+    expect(nil.to_tempus.data).to eq(0)
+  end
+end
